Moved Firebase _const to firebase module
