+++ src/auto_flutter_dig/task/config/firebase.py
@@ -3,7 +3,7 @@ from typing import Optional
 from ...core.utils import _Dict, _Enum, _If
 from ...model.argument.option import LongOption, LongOptionWithValue
 from ...model.platform import Platform
-from ..firebase._const import FIREBASE_PROJECT_APP_ID_KEY
+from ...module.firebase.model._const import FIREBASE_PROJECT_APP_ID_KEY
 from ._base import *
